+++ strum_macros/src/helpers/mod.rs
@@ -1,11 +1,9 @@
-pub use self::case_style::CaseStyleHelpers;
-pub use self::snakify::snakify;
+pub use self::case_style::{CaseStyleHelpers, snakify};
 pub use self::type_props::HasTypeProperties;
 pub use self::variant_props::HasStrumVariantProperties;
 
 pub mod case_style;
 mod metadata;
-mod snakify;
 pub mod type_props;
 pub mod variant_props;
